requirements.txt
streamlit>=1.40.0
anthropic>=0.42.0
openpyxl>=3.1.0
markdown>=3.7
weasyprint>=62.0
python-dotenv>=1.0.0
tenacity>=8.0.0
